+++ crates/miden-lib/src/account/interface/test.rs
@@ -15,7 +15,6 @@ use miden_objects::note::{
     NoteInputs,
     NoteMetadata,
     NoteRecipient,
-    NoteScript,
     NoteTag,
     NoteType,
 };
@@ -31,6 +30,7 @@ use crate::account::interface::{AccountInterface, NoteAccountCompatibility};
 use crate::account::wallets::BasicWallet;
 use crate::note::{create_p2id_note, create_p2ide_note, create_swap_note};
 use crate::transaction::TransactionKernel;
+use crate::utils::ScriptBuilder;
 
 // DEFAULT NOTES
 // ================================================================================================
@@ -279,9 +279,7 @@ fn test_basic_wallet_custom_notes() {
             end
         end
     ";
-    let note_script =
-        NoteScript::compile(compatible_source_code, TransactionKernel::testing_assembler())
-            .unwrap();
+    let note_script = ScriptBuilder::default().compile_note_script(compatible_source_code).unwrap();
     let recipient = NoteRecipient::new(serial_num, note_script, NoteInputs::default());
     let compatible_custom_note = Note::new(vault.clone(), metadata, recipient);
     assert_eq!(
@@ -310,8 +308,7 @@ fn test_basic_wallet_custom_notes() {
         end
     ";
     let note_script =
-        NoteScript::compile(incompatible_source_code, TransactionKernel::testing_assembler())
-            .unwrap();
+        ScriptBuilder::default().compile_note_script(incompatible_source_code).unwrap();
     let recipient = NoteRecipient::new(serial_num, note_script, NoteInputs::default());
     let incompatible_custom_note = Note::new(vault, metadata, recipient);
     assert_eq!(
@@ -371,9 +368,7 @@ fn test_basic_fungible_faucet_custom_notes() {
             end
         end
     ";
-    let note_script =
-        NoteScript::compile(compatible_source_code, TransactionKernel::testing_assembler())
-            .unwrap();
+    let note_script = ScriptBuilder::default().compile_note_script(compatible_source_code).unwrap();
     let recipient = NoteRecipient::new(serial_num, note_script, NoteInputs::default());
     let compatible_custom_note = Note::new(vault.clone(), metadata, recipient);
     assert_eq!(
@@ -404,8 +399,7 @@ fn test_basic_fungible_faucet_custom_notes() {
         end
     ";
     let note_script =
-        NoteScript::compile(incompatible_source_code, TransactionKernel::testing_assembler())
-            .unwrap();
+        ScriptBuilder::default().compile_note_script(incompatible_source_code).unwrap();
     let recipient = NoteRecipient::new(serial_num, note_script, NoteInputs::default());
     let incompatible_custom_note = Note::new(vault, metadata, recipient);
     assert_eq!(
@@ -487,13 +481,11 @@ fn test_custom_account_custom_notes() {
             end
         end
     ";
-    let note_script = NoteScript::compile(
-        compatible_source_code,
-        TransactionKernel::testing_assembler()
-            .with_dynamic_library(account_component.library())
-            .unwrap(),
-    )
-    .unwrap();
+    let note_script = ScriptBuilder::default()
+        .with_dynamically_linked_library(account_component.library())
+        .unwrap()
+        .compile_note_script(compatible_source_code)
+        .unwrap();
     let recipient = NoteRecipient::new(serial_num, note_script, NoteInputs::default());
     let compatible_custom_note = Note::new(vault.clone(), metadata, recipient);
     assert_eq!(
@@ -516,13 +508,11 @@ fn test_custom_account_custom_notes() {
             end
         end
     ";
-    let note_script = NoteScript::compile(
-        incompatible_source_code,
-        TransactionKernel::testing_assembler()
-            .with_dynamic_library(account_component.library())
-            .unwrap(),
-    )
-    .unwrap();
+    let note_script = ScriptBuilder::default()
+        .with_dynamically_linked_library(account_component.library())
+        .unwrap()
+        .compile_note_script(incompatible_source_code)
+        .unwrap();
     let recipient = NoteRecipient::new(serial_num, note_script, NoteInputs::default());
     let incompatible_custom_note = Note::new(vault, metadata, recipient);
     assert_eq!(
@@ -612,13 +602,11 @@ fn test_custom_account_multiple_components_custom_notes() {
             end
         end
     ";
-    let note_script = NoteScript::compile(
-        compatible_source_code,
-        TransactionKernel::testing_assembler()
-            .with_dynamic_library(custom_component.library())
-            .unwrap(),
-    )
-    .unwrap();
+    let note_script = ScriptBuilder::default()
+        .with_dynamically_linked_library(custom_component.library())
+        .unwrap()
+        .compile_note_script(compatible_source_code)
+        .unwrap();
     let recipient = NoteRecipient::new(serial_num, note_script, NoteInputs::default());
     let compatible_custom_note = Note::new(vault.clone(), metadata, recipient);
     assert_eq!(
@@ -653,13 +641,11 @@ fn test_custom_account_multiple_components_custom_notes() {
             end
         end
     ";
-    let note_script = NoteScript::compile(
-        incompatible_source_code,
-        TransactionKernel::testing_assembler()
-            .with_dynamic_library(custom_component.library())
-            .unwrap(),
-    )
-    .unwrap();
+    let note_script = ScriptBuilder::default()
+        .with_dynamically_linked_library(custom_component.library())
+        .unwrap()
+        .compile_note_script(incompatible_source_code)
+        .unwrap();
     let recipient = NoteRecipient::new(serial_num, note_script, NoteInputs::default());
     let incompatible_custom_note = Note::new(vault.clone(), metadata, recipient);
     assert_eq!(