+++ crates/miden-testing/src/kernel_tests/tx/test_fpi.rs
@@ -19,6 +19,7 @@ use miden_lib::transaction::memory::{
     NUM_ACCT_PROCEDURES_OFFSET,
     NUM_ACCT_STORAGE_SLOTS_OFFSET,
 };
+use miden_lib::utils::ScriptBuilder;
 use miden_objects::FieldElement;
 use miden_objects::account::{
     Account,
@@ -32,7 +33,7 @@ use miden_objects::account::{
 };
 use miden_objects::testing::account_component::AccountMockComponent;
 use miden_objects::testing::storage::STORAGE_LEAVES_2;
-use miden_objects::transaction::{AccountInputs, TransactionScript};
+use miden_objects::transaction::AccountInputs;
 use miden_tx::TransactionExecutorError;
 use rand::{Rng, SeedableRng};
 use rand_chacha::ChaCha20Rng;
@@ -57,7 +58,7 @@ fn test_fpi_memory() -> anyhow::Result<()> {
         use.miden::account
 
         export.get_item_foreign
-            # make this foreign procedure unique to make sure that we invoke the procedure of the 
+            # make this foreign procedure unique to make sure that we invoke the procedure of the
             # foreign account, not the native one
             push.1 drop
             exec.account::get_item
@@ -67,7 +68,7 @@ fn test_fpi_memory() -> anyhow::Result<()> {
         end
 
         export.get_map_item_foreign
-            # make this foreign procedure unique to make sure that we invoke the procedure of the 
+            # make this foreign procedure unique to make sure that we invoke the procedure of the
             # foreign account, not the native one
             push.2 drop
             exec.account::get_map_item
@@ -120,7 +121,7 @@ fn test_fpi_memory() -> anyhow::Result<()> {
     let code = format!(
         "
         use.std::sys
-        
+
         use.$kernel::prologue
         use.miden::tx
 
@@ -134,7 +135,7 @@ fn test_fpi_memory() -> anyhow::Result<()> {
             # push the index of desired storage item
             push.0
 
-            # get the hash of the `get_item_foreign` procedure of the foreign account 
+            # get the hash of the `get_item_foreign` procedure of the foreign account
             push.{get_item_foreign_hash}
 
             # push the foreign account ID
@@ -241,7 +242,7 @@ fn test_fpi_memory() -> anyhow::Result<()> {
             # push the index of desired storage item
             push.0
 
-            # get the hash of the `get_item_foreign` procedure of the foreign account 
+            # get the hash of the `get_item_foreign` procedure of the foreign account
             push.{get_item_foreign_hash}
 
             # push the foreign account ID
@@ -259,7 +260,7 @@ fn test_fpi_memory() -> anyhow::Result<()> {
             # push the index of desired storage item
             push.0
 
-            # get the hash of the `get_item_foreign` procedure of the foreign account 
+            # get the hash of the `get_item_foreign` procedure of the foreign account
             push.{get_item_foreign_hash}
 
             # push the foreign account ID
@@ -303,7 +304,7 @@ fn test_fpi_memory_two_accounts() -> anyhow::Result<()> {
         use.miden::account
 
         export.get_item_foreign_1
-            # make this foreign procedure unique to make sure that we invoke the procedure of the 
+            # make this foreign procedure unique to make sure that we invoke the procedure of the
             # foreign account, not the native one
             push.1 drop
             exec.account::get_item
@@ -316,7 +317,7 @@ fn test_fpi_memory_two_accounts() -> anyhow::Result<()> {
         use.miden::account
 
         export.get_item_foreign_2
-            # make this foreign procedure unique to make sure that we invoke the procedure of the 
+            # make this foreign procedure unique to make sure that we invoke the procedure of the
             # foreign account, not the native one
             push.2 drop
             exec.account::get_item
@@ -395,7 +396,7 @@ fn test_fpi_memory_two_accounts() -> anyhow::Result<()> {
         begin
             exec.prologue::prepare_transaction
 
-            ### Get the storage item at index 0 from the first account 
+            ### Get the storage item at index 0 from the first account
             # pad the stack for the `execute_foreign_procedure` execution
             padw padw padw push.0.0
             # => [pad(14)]
@@ -413,7 +414,7 @@ fn test_fpi_memory_two_accounts() -> anyhow::Result<()> {
             exec.tx::execute_foreign_procedure dropw
             # => []
 
-            ### Get the storage item at index 0 from the second account 
+            ### Get the storage item at index 0 from the second account
             # pad the stack for the `execute_foreign_procedure` execution
             padw padw padw push.0.0
             # => [pad(14)]
@@ -518,7 +519,7 @@ fn test_fpi_execute_foreign_procedure() -> anyhow::Result<()> {
         use.miden::account
 
         export.get_item_foreign
-            # make this foreign procedure unique to make sure that we invoke the procedure of the 
+            # make this foreign procedure unique to make sure that we invoke the procedure of the
             # foreign account, not the native one
             push.1 drop
             exec.account::get_item
@@ -528,7 +529,7 @@ fn test_fpi_execute_foreign_procedure() -> anyhow::Result<()> {
         end
 
         export.get_map_item_foreign
-            # make this foreign procedure unique to make sure that we invoke the procedure of the 
+            # make this foreign procedure unique to make sure that we invoke the procedure of the
             # foreign account, not the native one
             push.2 drop
             exec.account::get_map_item
@@ -627,7 +628,7 @@ fn test_fpi_execute_foreign_procedure() -> anyhow::Result<()> {
         map_key = STORAGE_LEAVES_2[0].0,
     );
 
-    let tx_script = TransactionScript::compile(code, TransactionKernel::testing_assembler())?;
+    let tx_script = ScriptBuilder::default().compile_tx_script(code)?;
 
     let foreign_account_inputs = mock_chain
         .get_foreign_account_inputs(foreign_account.id())
@@ -661,7 +662,7 @@ fn test_nested_fpi_cyclic_invocation() -> anyhow::Result<()> {
     let second_foreign_account_code_source = r#"
         use.miden::tx
         use.miden::account
-        
+
         use.std::sys
 
         export.second_account_foreign_proc
@@ -686,7 +687,7 @@ fn test_nested_fpi_cyclic_invocation() -> anyhow::Result<()> {
             # make sure that the resulting value equals 5
             dup push.5 assert_eq.err="value should have been 5"
 
-            # get the first element of the 0'th storage slot (it should be 1) and add it to the 
+            # get the first element of the 0'th storage slot (it should be 1) and add it to the
             # obtained foreign value.
             push.0 exec.account::get_item drop drop drop
             add
@@ -734,7 +735,7 @@ fn test_nested_fpi_cyclic_invocation() -> anyhow::Result<()> {
             exec.tx::execute_foreign_procedure
             # => [storage_value]
 
-            # get the second element of the 0'th storage slot (it should be 2) and add it to the 
+            # get the second element of the 0'th storage slot (it should be 2) and add it to the
             # obtained foreign value.
             push.0 exec.account::get_item drop drop swap drop
             add
@@ -746,7 +747,7 @@ fn test_nested_fpi_cyclic_invocation() -> anyhow::Result<()> {
         end
 
         export.get_item_foreign
-            # make this foreign procedure unique to make sure that we invoke the procedure of the 
+            # make this foreign procedure unique to make sure that we invoke the procedure of the
             # foreign account, not the native one
             push.1 drop
             exec.account::get_item
@@ -830,13 +831,13 @@ fn test_nested_fpi_cyclic_invocation() -> anyhow::Result<()> {
 
             exec.tx::execute_foreign_procedure
             # => [storage_value]
-            
+
             # add 10 to the returning value
             add.10
 
             # assert that the resulting value equals 18
             push.18 assert_eq.err="sum should be 18"
-            # => []        
+            # => []
 
             exec.sys::truncate_stack
         end
@@ -846,10 +847,7 @@ fn test_nested_fpi_cyclic_invocation() -> anyhow::Result<()> {
         first_account_foreign_proc_hash = first_foreign_account.code().procedures()[1].mast_root(),
     );
 
-    let tx_script = TransactionScript::compile(
-        code,
-        TransactionKernel::testing_assembler().with_debug_mode(true),
-    )?;
+    let tx_script = ScriptBuilder::default().compile_tx_script(code)?;
 
     let tx_context = mock_chain
         .build_tx_context(native_account.id(), &[], &[])
@@ -1017,11 +1015,9 @@ fn test_nested_fpi_stack_overflow() {
                 foreign_suffix = foreign_accounts.last().unwrap().id().suffix(),
             );
 
-            let tx_script = TransactionScript::compile(
-                code,
-                TransactionKernel::testing_assembler().with_debug_mode(true),
-            )
-            .unwrap();
+
+
+                let tx_script = ScriptBuilder::default().compile_tx_script(code).unwrap();
 
             let tx_context = mock_chain
                 .build_tx_context(native_account.id(), &[], &[])
@@ -1138,10 +1134,7 @@ fn test_nested_fpi_native_account_invocation() -> anyhow::Result<()> {
         first_account_foreign_proc_hash = foreign_account.code().procedures()[1].mast_root(),
     );
 
-    let tx_script = TransactionScript::compile(
-        code,
-        TransactionKernel::testing_assembler().with_debug_mode(true),
-    )?;
+    let tx_script = ScriptBuilder::default().compile_tx_script(code)?;
 
     let tx_context = mock_chain
         .build_tx_context(native_account.id(), &[], &[])