+++ crates/miden-testing/tests/scripts/p2id.rs
@@ -1,6 +1,6 @@
 use miden_lib::errors::note_script_errors::ERR_P2ID_TARGET_ACCT_MISMATCH;
 use miden_lib::note::create_p2id_note;
-use miden_lib::transaction::TransactionKernel;
+use miden_lib::utils::ScriptBuilder;
 use miden_objects::account::Account;
 use miden_objects::asset::{Asset, AssetVault, FungibleAsset};
 use miden_objects::crypto::rand::RpoRandomCoin;
@@ -12,7 +12,7 @@ use miden_objects::testing::account_id::{
     ACCOUNT_ID_REGULAR_PUBLIC_ACCOUNT_IMMUTABLE_CODE_2,
     ACCOUNT_ID_SENDER,
 };
-use miden_objects::transaction::{OutputNote, TransactionScript};
+use miden_objects::transaction::OutputNote;
 use miden_objects::{Felt, Word};
 use miden_testing::{Auth, MockChain};
 use miden_tx::utils::word_to_masm_push_string;
@@ -257,8 +257,7 @@ fn test_create_consume_multiple_notes() -> anyhow::Result<()> {
         note_execution_hint_2 = Felt::from(output_note_2.metadata().execution_hint())
     );
 
-    let tx_script =
-        TransactionScript::compile(tx_script_src, TransactionKernel::testing_assembler())?;
+    let tx_script = ScriptBuilder::default().compile_tx_script(tx_script_src)?;
 
     let tx_context = mock_chain
         .build_tx_context(account.id(), &[input_note_1.id(), input_note_2.id()], &[])?